[#20] Move AreaChange to its own module.
The area change related source terms have been moved from geometry to
models/area_change.py. This commit includes a couple minor formatting
and typing fixes.
